updating unit test runner, removing unused unit test files
