+++ package.json
@@ -29,11 +29,11 @@
     "start": "webpack-dev-server --config webpack.config.dev.js",
     "start-gulp": "gulp serve watch",
     "release": "gulp release",
-    "test": "gulp test",
-    "unit": "gulp unit",
+    "test": "npm run test && npm run e2e",
+    "unit": "karma start test/karma.conf.js",
     "e2e": "protractor test/protractor.conf.js",
     "report-coverage": "codecov",
-    "lint": "eslint src/**/*.js src/components/**/*.js src/components/menu/**/*.js src/*.js test/*.js config.js gulpfile.babel.js --fix",
+    "lint": "eslint src/**/*.js src/components/**/*.js src/components/menu/**/*.js src/*.js test/*.js config.js gulpfile.babel.js",
     "stylelint": "stylelint src/dataTable.less src/themes/*.less src/components/menu/menu.less src/components/popover/popover.less"
   },
   "peerDependencies": {
@@ -46,18 +46,18 @@
     "babel-eslint": "^7.0.0",
     "babel-loader": "^7.0.0",
     "babel-plugin-external-helpers": "^6.5.0",
+    "babel-plugin-rewire": "^1.0.0",
+    "babel-plugin-root-import": "^5.1.0",
     "babel-plugin-transform-es2015-modules-amd": "^6.8.0",
     "babel-plugin-transform-es2015-modules-commonjs": "^6.8.0",
     "babel-plugin-transform-es2015-modules-umd": "^6.8.0",
     "babel-plugin-transform-object-assign": "^6.8.0",
     "babel-preset-es2015": "^6.9.0",
     "babel-preset-stage-0": "^6.5.0",
-    "babel-root-import": "^4.1.3",
-    "babelify": "7.3.0",
+    "babelify": "^7.3.0",
     "bardjs": "^0.1.8",
-    "brfs": "^1.4.3",
     "browser-sync": "^2.7.2",
-    "browserify": "^13.1.1",
+    "browserify": "^14.0.0",
     "browserify-istanbul": "^2.0.0",
     "codecov": "^1.0.1",
     "css-loader": "^0.28.1",
@@ -86,14 +86,14 @@
     "isparta": "^4.0.0",
     "istanbul": "^0.4.5",
     "jasmine-core": "^2.5.2",
-    "karma": "~1.5.0",
-    "karma-angular": "0.0.6",
+    "karma": "^1.4.1",
     "karma-babel-preprocessor": "^6.0.1",
-    "karma-browserify": "^5.1.0",
+    "karma-browserify": "^5.1.1",
     "karma-coverage": "^1.1.1",
-    "karma-jasmine": "^1.0.2",
-    "karma-phantomjs-launcher": "^1.0.2",
-    "karma-sauce-launcher": "^1.1.0",
+    "karma-jasmine": "^1.1.0",
+    "karma-mocha-reporter": "^2.2.3",
+    "karma-phantomjs-launcher": "1.0.2",
+    "karma-phantomjs-shim": "^1.4.0",
     "karma-sinon": "^1.0.5",
     "less-loader": "^4.0.3",
     "phantomjs-prebuilt": "^2.1.7",
