--- src/components/body/GroupRowController.spec.js
@@ -1,5 +1,0 @@
-// import GroupRowController from './GroupRowController';
-
-describe('GroupRowController', () => {
-  //
-});
