--- src/components/body/RowController.spec.js
@@ -1,5 +1,0 @@
-// import RowController from './RowController';
-
-describe('RowController', () => {
-  //
-});
