--- src/components/body/SelectionController.spec.js
@@ -1,5 +1,0 @@
-// import SelectionController from './SelectionController';
-
-describe('SelectionController', () => {
-  //
-});
