--- src/components/body/StyleTranslator.spec.js
@@ -1,5 +1,0 @@
-// import StyleTranslator from './StyleTranslator';
-
-describe('StyleTranslator', () => {
-  //
-});
