--- src/components/footer/FooterController.spec.js
@@ -1,5 +1,0 @@
-// import FooterController from './FooterController';
-
-describe('FooterController', () => {
-  //
-});
