--- src/components/header/HeaderCellController.spec.js
@@ -1,5 +1,0 @@
-// import HeaderCellController from './HeaderCellController';
-
-describe('HeaderCellController', () => {
-  //
-});
