--- src/components/menu/MenuController.spec.js
@@ -1,5 +1,0 @@
-// import MenuController from './MenuController';
-
-describe('MenuController', () => {
-  //
-});
