--- src/components/menu/dropdown/DropdownController.spec.js
@@ -1,5 +1,0 @@
-// import DropdownController from './DropdownController';
-
-describe('DropdownController', () => {
-  //
-});
